move support to trial.py
