+++ e2e_tests/tests/trial/support-bundle.py
@@ -1,0 +1,2 @@
+import experiment 
+
